requirements.txt
streamlit
openai
pandas
altair
python-dotenv